database changes in acc to frontend

+++ database/migrations/2024_11_23_114258_create_user_stats_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->integer('protein')->default(0);
             $table->integer('fat')->default(0);
             $table->integer('carbs')->default(0);
-            $table->decimal('liquid_intake', 5, 2)->default(0.0);
+            $table->integer('liquid_intake')->default(0);
 
             $table->timestamps();
